Move boundary list functionality into parent namedBoundaryMenu
