+++ app/src/components/map/Tools/ToolTypes/Nav/JumpToTrip.tsx
@@ -1,7 +1,5 @@
 import { Capacitor } from '@capacitor/core';
-import { Box, Button, ListItem, ListItemIcon, ListItemText, Typography } from '@mui/material';
-import ExploreIcon from '@mui/icons-material/Explore';
-import { ListItemButton } from '@mui/material';
+import { Button, ListItem, ListItemIcon, ListItemText, Typography } from '@mui/material';
 import { DatabaseContext } from 'contexts/DatabaseContext';
 import { useDataAccess } from 'hooks/useDataAccess';
 import L from 'leaflet';
@@ -13,7 +11,6 @@ import { FlyToAndFadeItemTransitionType, IFlyToAndFadeItem, useFlyToAndFadeConte
 
 import CheckIcon from '@mui/icons-material/Check';
 import DeleteIcon from '@mui/icons-material/Delete';
-import EditTools from '../Data/EditTools';
 /*
 
 - [ ] add ui button to let user add shapes
@@ -21,13 +18,6 @@ import EditTools from '../Data/EditTools';
 - [ ] or view existing in list
 - [ ] 
 */
-
-interface Boundary {
-  id: number,
-  name: string,
-  geos: [],
-  server_id: number
-}
 
 export const JumpToTrip = (props) => {
   // style
@@ -44,10 +34,9 @@ export const JumpToTrip = (props) => {
 
   // initial setup & events to block:
   useEffect(() => {
-    L.DomEvent.disableClickPropagation(divRef?.current);
-    L.DomEvent.disableScrollPropagation(divRef?.current);
-    getTripGeosAndInitialPosition();
-    getBoundaries();
+    // L.DomEvent.disableClickPropagation(divRef?.current);
+    // L.DomEvent.disableScrollPropagation(divRef?.current);
+    // getTripGeosAndInitialPosition();
   }, []);
 
   // What the button cycles through.
@@ -55,12 +44,10 @@ export const JumpToTrip = (props) => {
   const [IFlyToAndFadeItems, setIFlyToAndFadeItems] = useState<Array<IFlyToAndFadeItem>>([]);
   const [index, setIndex] = useState<number>(0);
   const [edit, setEdit] = useState(false);
-  const [boundaries, setBoundaries] = useState<Boundary[]>([]);
-  const [idCount, setIdCount] = useState(0);
 
   // map Event subcriptions:
   const map = useMapEvent('dragend', () => {
-    getTripGeosAndInitialPosition();
+    // getTripGeosAndInitialPosition();
   });
 
   //onclick:
@@ -81,175 +68,105 @@ export const JumpToTrip = (props) => {
   }, [index]);
 
   // can be replaced with a menu (later):
-  const getTripGeosAndInitialPosition = async () => {
-    let tripObjects;
-    //mobile only
-    if (Capacitor.getPlatform() === 'ios' || Capacitor.getPlatform() === 'android') {
-      const queryResults = await dataAccess.getTrips();
-      if (!queryResults.length) {
-        return;
-      }
-      tripObjects = queryResults.map((rawRecord) => {
-        return JSON.parse(rawRecord.json);
-      });
-    } else {
-      tripObjects = [
-        {
-          id: 1,
-          name: 'trip a',
-          geometry: [
-            {
-              type: 'Feature',
-              properties: {},
-              geometry: {
-                type: 'Polygon',
-                coordinates: [
-                  [
-                    [-126.826171875, 51.876490970614775],
-                    [-123.70605468750001, 51.876490970614775],
-                    [-123.70605468750001, 53.68369534495075],
-                    [-126.826171875, 53.68369534495075],
-                    [-126.826171875, 51.876490970614775]
-                  ]
-                ]
-              }
-            }
-          ]
-        }
-      ];
-    }
+  // const getTripGeosAndInitialPosition = async () => {
+  //   let tripObjects;
+  //   //mobile only
+  //   if (Capacitor.getPlatform() === 'ios' || Capacitor.getPlatform() === 'android') {
+  //     const queryResults = await dataAccess.getTrips();
+  //     if (!queryResults.length) {
+  //       return;
+  //     }
+  //     tripObjects = queryResults.map((rawRecord) => {
+  //       return JSON.parse(rawRecord.json);
+  //     });
+  //   } else {
+  //     tripObjects = [
+  //       {
+  //         id: 1,
+  //         name: 'trip a',
+  //         geometry: [
+  //           {
+  //             type: 'Feature',
+  //             properties: {},
+  //             geometry: {
+  //               type: 'Polygon',
+  //               coordinates: [
+  //                 [
+  //                   [-126.826171875, 51.876490970614775],
+  //                   [-123.70605468750001, 51.876490970614775],
+  //                   [-123.70605468750001, 53.68369534495075],
+  //                   [-126.826171875, 53.68369534495075],
+  //                   [-126.826171875, 51.876490970614775]
+  //                 ]
+  //               ]
+  //             }
+  //           }
+  //         ]
+  //       }
+  //     ];
+  //   }
 
-    let items = new Array<IFlyToAndFadeItem>();
+  //   let items = new Array<IFlyToAndFadeItem>();
 
-    //add current position as bounds to zoom to
-    items.push({
-      name: 'Original Position',
-      bounds: map.getBounds(),
-      colour: 'red',
-      transitionType: FlyToAndFadeItemTransitionType.zoomToBounds
-    });
-    //then add trips as geometries to show
-    for (const trip of tripObjects.sort((a, b) => (a.id < b.id ? 1 : -1))) {
-      if (trip.geometry.length > 0) {
-        items.push({
-          name: 'TRIP: ' + trip.name,
-          geometries: trip.geometry,
-          colour: 'red',
-          transitionType: FlyToAndFadeItemTransitionType.zoomToGeometries
-        });
-      }
-    }
+  //   //add current position as bounds to zoom to
+  //   items.push({
+  //     name: 'Original Position',
+  //     bounds: map.getBounds(),
+  //     colour: 'red',
+  //     transitionType: FlyToAndFadeItemTransitionType.zoomToBounds
+  //   });
+  //   //then add trips as geometries to show
+  //   for (const trip of tripObjects.sort((a, b) => (a.id < b.id ? 1 : -1))) {
+  //     if (trip.geometry.length > 0) {
+  //       items.push({
+  //         name: 'TRIP: ' + trip.name,
+  //         geometries: trip.geometry,
+  //         colour: 'red',
+  //         transitionType: FlyToAndFadeItemTransitionType.zoomToGeometries
+  //       });
+  //     }
+  //   }
 
-    setIFlyToAndFadeItems([...items]);
-  };
+  //   setIFlyToAndFadeItems([...items]);
+  // };
 
-  const setBoundaryIdCount = (() => {
-    if (boundaries && boundaries.length > 0) {
-      //ensures id is not repeated on client side
-      const max = Math.max(...boundaries.map(b => b.id));
-      setIdCount(max + 1);
-    }
-  });
-
-  useEffect(() => {
-    setBoundaryIdCount();
-  }, [boundaries]);
-
-  const getBoundaries = async () => {
-    const results = await dataAccess.getBoundaries();
-    if (results) {
-      setBoundaries(results);
-    }
-  };
-
-  const createBoundary = (() => {
-    const dowe = window.confirm('Create new named boundary?');
-    if (dowe) {
-      const name = prompt('Name:');
-      setEdit(true);
-
-      const tempBoundary: Boundary = {
-        id: idCount,
-        name: name,
-        geos: [],
-        server_id: null
-      };
-
-      dataAccess.addBoundary(tempBoundary);
-      setBoundaries([...boundaries, tempBoundary]);
-    }
-  });
-
-  const deleteBoundary = async (id: number) => {
-    await dataAccess.deleteBoundary(id);
-    getBoundaries();
-  }
 
   return (
-    <Box>
-      <ListItem disableGutters>
-        <ListItemButton
-          onClick={createBoundary}
-          ref={divRef}
-          aria-label="Jump To Location"
-          style={{ padding: 10, borderBottomLeftRadius: 5, borderBottomRightRadius: 5 }}>
-          <ListItemIcon>
-            <ExploreIcon />
-          </ListItemIcon>
-          <ListItemText>
-            <Typography className={toolClass.Font}>New Boundary</Typography>
-          </ListItemText>
-        </ListItemButton>
-      </ListItem>
-
-      {boundaries.map((boundary, index) => (
-        <ListItem
-          key={index}
-          onClick={() => {
-            jump();
-          }}
-          disableGutters>
-          <ListItemText>
-            <Typography className={toolClass.Font}>{boundary.name}</Typography>
-          </ListItemText>
-          <ListItemIcon>
-            <CheckIcon />
-          </ListItemIcon>
-          <ListItemIcon>
-            <Button onClick={() => deleteBoundary(boundary.id)}>
-              <DeleteIcon />
-            </Button>
-          </ListItemIcon>
-        </ListItem>
-      ))}
-      {/* <ListItem
-        onClick={() => {
-          jump();
-        }}
-        disableGutters>
-        <ListItemText>
-          <Typography className={toolClass.Font}>Sunny infested areas</Typography>
-        </ListItemText>
-        <ListItemIcon>
-          <CheckIcon />
-        </ListItemIcon>
-      </ListItem> */}
-      {/* <ListItem disableGutters>
-        <ListItemText>
-          <Typography className={toolClass.Font}>Scenic infested areas</Typography>
-        </ListItemText>
-      </ListItem> */}
-      {/*
-        <>
-          edit?
-          <FeatureGroup>
-            <EditTools />
-          </FeatureGroup>
-          : <></>
-        </>
-      */}
-    </Box>
+    <ListItem
+      onClick={() => {
+        jump();
+      }}
+      disableGutters>
+      <ListItemText>
+        <Typography className={toolClass.Font}>{props.name}</Typography>
+      </ListItemText>
+      <ListItemIcon>
+        <CheckIcon />
+      </ListItemIcon>
+      <ListItemIcon>
+        <Button onClick={() => props.deleteBoundary(props.id)}>
+          <DeleteIcon />
+        </Button>
+      </ListItemIcon>
+    </ListItem>
   );
 };
 
 export default JumpToTrip;
+
+
+/* 
+just for reference until a later commit 
+
+<ListItem
+  onClick={() => {
+    jump();
+  }}
+  disableGutters>
+  <ListItemText>
+    <Typography className={toolClass.Font}>Sunny infested areas</Typography>
+  </ListItemText>
+  <ListItemIcon>
+    <CheckIcon />
+  </ListItemIcon>
+</ListItem> */
